Update 09-portrayal.adoc
Per Topi's comment during PC Discussion (29 June 2023), I investigated and found the following: (a) When our depth sign convention was pos-up, 315 degrees was our azimuthal angle. (b) The azimuthal angle (erroneously) remained 315 despite our change to pos-down. (c) Topi demonstrated the necessity of the change cogently in an email, and he gave further evidence in a blog post (https://ramblemaps.com/why-does-sunlight-come-from-north). As such, it is right to change this value to 135 degrees. ∎