Removed table for exchange catalogue params
Heading was previously (erroneously) named S102_ExchangeCatalogue.

The only modified parameter wrt S-100 Ed. 5.2.0 was a restriction on the multiplicity of identifier (restricted to 1 vice 0..1). However, it seems improper to force "1" upon the entire exchange set when S-100 intends to allow other products to omit this value. With a change to 0..1, no parameters are modified, and thus the table can now be omitted.