Modified Figure for QualityOfBathymetryCoverage
This figure is exactly the same as the nozero version, but the name of this one is what is referenced at present in Section 10. N.B.: the two figures differed in earlier versions. QualityOfSurvey was present in this file while QualityOfSurveyCoverage was present in the nozero version. Neither file featured level 0.