Changed link from S-100 Reference to S-102 Image
the image S-102 uses is one and the same as what was previously being referred to. It is appropriate in this case to directly use the image (to enable ready visualization of the model without having to open another spec/standard), and so, instead of referring out to S-100, I'm changing the reference to the figure contained earlier in this section.